Reorganized to examples
