+++ docs/contents.rst
@@ -9,7 +9,7 @@ PHPBench is a benchmarking framework for PHP. Find it on Github_.
    installing
    quick-start
    guides
-   recipes
+   examples
    reference
    extensions
    customisation
