+++ docs/examples.rst
@@ -1,9 +1,10 @@
-Recipes
--------
+Examples
+--------
 
 This section aims to provide a collection of benchmarking scenarios with examples.
 
 .. toctree::
    :maxdepth: 2
 
-   recipes/file_input
+   examples/file_input
+   examples/hashing
