+++ examples/Benchmark/FileInput/FileInputBench.php
