+++ examples/Recipes/report/bar_chart_by_parameter.json
@@ -1,0 +1,26 @@
+{
+    "$schema": "./../../../phpbench.schema.json",
+    "runner.bootstrap": "../../../vendor/autoload.php",
+    "runner.path": "../../../examples/Benchmark/Micro/HashingBench.php",
+    "report.outputs": {
+        "recipe": {
+            "renderer": "html",
+            "path": "examples/Recipes/report/report.html"
+        }
+    },
+    "report.generators": {
+        "hashing": {
+            "generator": "component",
+            "filter": "benchmark_name = 'HashingBench'",
+            "components": [
+                {
+                    "component": "bar_chart_aggregate",
+                    "x_partition": "variant_params['algo']",
+                    "bar_partition": "variant_params['size']",
+                    "y_expr": "mode(partition['result_time_avg']) as time",
+                    "y_axes_label": "yValue as time"
+                }
+            ]
+        }
+    }
+}
