+++ examples/Recipes/report/report.html
@@ -1,0 +1,62 @@
+<html>
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>PHPBench Report</title>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet">
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js"></script>
+        <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
+    </head>
+    <body class="container">
+                <h1>PHPBench Report</h1>
+        <hr/>
+                    <div>
+                                <div id="phpbench-component-4" width="400" height="100"></div>
+<script>
+var data = [
+        {
+        x: ["md5","sha256"],
+        name: '',
+        y: [0.5202348336594916,1.3133444227005877],
+        type: 'bar',
+            },
+    ]
+var layout = {
+    barmode: 'group',
+    yaxis: {
+        tickmode: 'array',
+        tickvals: [0,0.13133444227005878,0.26266888454011755,0.39400332681017636,0.5253377690802351,0.6566722113502939,0.7880066536203527,0.9193410958904115,1.0506755381604702,1.182009980430529,1.3133444227005877],
+        ticktext: ["0.000<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.131<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.263<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.394<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.525<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.657<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.788<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","0.919<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","1.051<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","1.182<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n","1.313<span style=\"color: #555; font-weight: bold;\">\u03bcs<\/span>\n"]    },
+    xaxis: {
+        tickmode: 'array',
+        tickvals: ["md5","sha256"],
+        ticktext: ["md5","sha256"],
+    }
+};
+Plotly.newPlot(document.getElementById('phpbench-component-4'), data, layout);
+</script>
+            </div>
+    </body>
+    <footer>
+        <center>
+        Generated by <a href="https://github.com/phpbench/phpbench">PHPBench</a>
+        </center>
+    </footer>
+    <script type="text/javascript">
+    /**
+     * Resize Plots when tab is activated
+     */
+    (function () {
+        var tabEls = document.querySelectorAll('button[data-bs-toggle="tab"]')
+        tabEls.forEach(function (tabEl) {
+            tabEl.addEventListener('shown.bs.tab', function (event) {
+console.log(event);
+                var docs = event.target.parentNode.parentNode.parentNode.querySelectorAll('.tab-pane.active .js-plotly-plot');
+                docs.forEach(function (plotEl) {
+                    Plotly.relayout(plotEl, {autosize: true});
+                });
+            });
+        });
+    })();
+    </script>
+</html>
